Updating the frontend

+++ database/migrations/2019_12_26_024019_create_expenses_table.php
@@ -17,7 +17,7 @@ class CreateExpensesTable extends Migration
             $table->bigIncrements('id');
             $table->unsignedBigInteger('category_id');
 
-            $table->bigInteger('amount');
+            $table->decimal('amount',8,2);
             
             $table->unsignedBigInteger('recepient_id')->nullable();
